src/engine/requirements.txt
keras==3.3.3
tensorflow==2.16.1
numpy==1.26.4
torch==2.3.0
pillow==10.3.0
scipy==1.13.1
fastapi==0.111.0
uvicorn==0.30.1
azure-storage-blob==12.20.0
python-dotenv==1.0.1
ultralytics==8.2.28